web/requirements.txt
Flask==3.0.0
pandas==2.1.1
Flask-SQLAlchemy==3.1.1
numpy<2.0.0
python-dotenv
requests
yfinance
xlrd
openpyxl
firebase-admin
gunicorn
